[feat] add l10n folder to separate l10n related files from common data
