+++ src/app/_components/profile/Profile.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import ProfileImage from "@images/profile.jpeg";
 import { defaultLang, SupportedLangs } from "../../types";
-import data from "../../../data";
+import data from "../../../data/l10n";
 
 type ProfileProps = {
   lang?: SupportedLangs;
