+++ src/data/l10n/index.ts
