+++ src/data/l10n/kr/about.json
