+++ src/data/l10n/kr/index.ts
